Bootcamp2/week15/requirements.txt
psycopg2-binary
pandas
boto3
python-dotenv